+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 5/26/2015
+- Enhaned `New-RSGroup` and `Set-RSGroup` to allow manipulation of reference groups.
+
 ## 2/25/2015
 - Added `Get-RSCollectionTasks` implementation to fetch data collection tasks.